sddd

+++ database/migrations/2021_05_14_000851_create_d_e_o_s_controllers_table.php
@@ -19,8 +19,9 @@ class CreateDEOSControllersTable extends Migration
             $table->text('name')->nullable();
             $table->string('ip_address')->nullable();
             $table->string('port_number')->nullable();
-            $table->float('longitude')->nullable();
-            $table->float('latitude')->nullable();
+            $table->double('longitude', 10, 4)->nullable();
+            $table->double('latitude', 10, 4)->nullable();
+
 
             $table->unsignedBigInteger('building_id')->nullable();
             $table->foreign('building_id')->references('id')->on('buildings')->onDelete('cascade');
